requirements.txt
pandas
streamlit
accelerate 
datasets 
pymongo 
sentence_transformers
-U transformers
geniusreader
lyricsgenius